Fix 2904686 https://sourceforge.net/tracker/?func=detail&aid=2904686&group_id=216267&atid=1036998
I made it so that origins and terminal names are automatically unique. 

Details:
I couldn't quite get the name resolution popup to work in this code path because of the UI architecture. 

Basically, there's a UI blocking thread waiting for this task to complete, and therefore I cannot show any new dialogs in the add origin/termination code path.

Ideally, there'd also be a way to rename origins/terminations out of band, but this functionality doesn't exist in NEFEnsemble yet.