Addresses (but doesn't fix) #409
This prevents Nengo from crashing after changing the # of neurons
in a population while interactive plots is simulating. However, most
of the graphs will still fail to reflect the change.

* in timelog.py an array was being concatenated with a list, which
isn't allowed. using extend instead.

* made setNodeCount synchronized so that the # of nodes in an ensmble
can be changed while an interactive plots session is simulating - kind
of hackey, and there is probably more that needs to be fixed, but
seems to address this issue.