Fix #407
The ensemble viewer will now update when the number of cells are
changed.

* UINEFEnsemble was not calling update on its viewer when the model
was changed. So overrode the modelUpdated function which fetches
the viewer and calls updateViewFromModel on it. Analogous to how the
UINetwork updates the NetworkViewer.

* Changed the function EnsembleViewer.updateViewFromModel to work
very similarly to how NetworkViewer.updateViewFromModel works. The
original version was giving synchronization errors.