Import template in startup-ui.py
I guess this was previously being done elsewhere? In some other .py
file? In any case, it should only be done in startup-ui.py, as there's
no need to import this otherwise. Really, these kinds of after-effects
should not happen as a result of import (it should be explicit),
but there's a lot of refactoring that should happen in here; this isn't
really that important. Anyway, Nengo runs properly for me now.