Move Java docs to Sphinx
The newly created `docs` directory contains these docs. The bulk
of the directory was created automatically with javasphinx-apidoc;
however, the output needed minor modifications to have the docs
build properly.

The next step will be to move the docs currently in
simulator-ui/docs to this new docs directory in order to keep all
of the Nengo 1.4 docs in one place.