Revamp documentation
I went through all of the ReST files in the docs to fix up the
ReST formatting, fix PEP8 in the code samples and demos, and
ensure that everything looks decent.

In addition to these changes, I added warnings to several pages
indicating that Nengo 1.4 is outdated so please upgrade to Nengo 2.